fix(logs essentials): actions are shown in alert details page
